parser/pyproject.toml
[project]
name = "parser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "minio>=7.2.0",
    "openai>=1.0.0",
    "openpyxl>=3.1.0",
    "opensearch-py>=2.4.0",
    "pdfplumber>=0.11.0",
    "pillow>=10.0.0",
    "psycopg2-binary>=2.9.0",
    "pypdfium2>=4.0.0",
    "python-docx>=1.1.0",
    "python-pptx>=1.0.0",
]
